requirements.txt
RPi.bme280
mysqlclient
